fix general linting errors
