+++ packages/next/cms/components/Viz/Table.js
@@ -1,5 +1,6 @@
 import React from "react";
 
+/** TODO */
 export default function Table() {
   return <div>Table</div>;
 }
