+++ packages/next/cms/components/sections/SubGrouping.js
@@ -1,3 +1,4 @@
+// eslint-disable-next-line no-unused-vars
 import React from "react";
 import Grouping from "./Grouping";
