+++ packages/next/cms/components/sections/Tabs.js
@@ -1,4 +1,6 @@
 import React from "react";
+
+/** TODO */
 export default function TabsSection() {
   return <div>TabsSection Section</div>;
 }
