+++ packages/next/cms/hooks/index.js
@@ -1,3 +1,6 @@
 export {default as useProfileSections} from "./use-profile-sections";
 export {default as useComparison} from "./use-comparison";
 export {default as useReportState} from "./use-report-state";
+export {default as useOnSelector} from "./use-on-selector";
+export {default as useOnSetVariables} from "./use-on-set-variables";
+export {default as useOnTabSelect} from "./use-on-tab-select";
